+++ docs/networks/mainnet/parameters.md
@@ -34,7 +34,7 @@ Developers can use the services of the following 3rd party providers:
 
 ## Add LUKSO Mainnet (MetaMask...)
 
-<AddNetworkButton/>
+<AddNetworkButton networkName="mainnet"/>
 
 You can add LUKSO as a custom network in your existing wallet (MetaMask, Rainbow, Coinbase Wallet, Trust Wallet, etc.) with the parameters below:
