+++ src/components/AddNetworkButton/index.js
@@ -1,23 +1,15 @@
 import React from 'react';
 
+import { LUKSO_NETWORK_CONFIGS } from '../../constants';
+
 import styles from './AddNetworkButton.module.scss';
 
-// TODO: we could use a global config file and fetch the info from there on all pages
-const LUKSO_NETWORK_CONFIGS = {
-  mainnet: {
-    chainId: '0x2A', // 42
-    chainName: 'LUKSO',
-    nativeCurrency: {
-      name: 'LYX',
-      symbol: 'LYX',
-      decimals: 18,
-    },
-    rpcUrls: ['https://rpc.lukso.gateway.fm'],
-    blockExplorerUrls: ['https://explorer.execution.mainnet.lukso.network'],
-  },
-};
-
-export default function AddNetworkButton() {
+/**
+ *
+ * @param {*} networkName: either mainnet or testnet
+ * @returns
+ */
+export default function AddNetworkButton({ networkName }) {
   const addNetwork = async () => {
     const ethereum = window.ethereum;
 
@@ -29,7 +21,7 @@ export default function AddNetworkButton() {
     try {
       await ethereum.request({
         method: 'wallet_switchEthereumChain',
-        params: [{ chainId: LUKSO_NETWORK_CONFIGS.mainnet.chainId }],
+        params: [{ chainId: LUKSO_NETWORK_CONFIGS[networkName].chainId }],
       });
       alert('Your extension is now connected to LUKSO network.');
     } catch (switchError) {
@@ -38,7 +30,7 @@ export default function AddNetworkButton() {
         try {
           await ethereum.request({
             method: 'wallet_addEthereumChain',
-            params: [LUKSO_NETWORK_CONFIGS.mainnet],
+            params: [LUKSO_NETWORK_CONFIGS[networkName]],
           });
         } catch (addError) {
           alert(addError.message);
@@ -51,7 +43,7 @@ export default function AddNetworkButton() {
 
   return (
     <button className={styles.button} onClick={addNetwork}>
-      Add LUKSO Mainnet
+      Add LUKSO {networkName}
     </button>
   );
 }
