+++ src/components/AddNetworkButtonTestnet/AddNetworkButton.module.scss
@@ -1,6 +1,6 @@
 .button {
   tab-size: 4;
-  transition: box-shadow .3s;
+  transition: box-shadow 0.3s;
 
   box-sizing: border-box;
   border: 0 solid #f2f2f2;
@@ -25,6 +25,6 @@
   color: white;
 
   &:hover {
-    box-shadow: 0 0 11px rgba(33,33,33,.7); 
+    box-shadow: 0 0 11px rgba(33, 33, 33, 0.7);
   }
 }
